Added Sound Effects to Puzzle Room
I added sound effects to the sculpture and walls that are moving. The sculpture sound effect is looping to allow the player to find it, where the wall audio starts when it starts moving and ends when it is done. Might have to play around with volume, but wanted to get it in for the demo.